--- a/mypy/util.py
+++ b/mypy/util.py
@@ -417,6 +417,20 @@ def is_sub_path(path1: str, path2: str) -> bool:
     return pathlib.Path(path2) in pathlib.Path(path1).parents
 
 
+if sys.platform == "linux" or sys.platform == "darwin":
+    def os_path_join(path: str, b: str) -> str:
+        # Based off of os.path.join, but simplified to str-only, 2 args and mypyc can compile it.
+        if b.startswith("/") or not path:
+            return b
+        elif path.endswith("/"):
+            return path + b
+        else:
+            return path + "/" + b
+else:
+    def os_path_join(a: str, p: str) -> str:
+        return os.path.join(a, p)
+
+
 def hard_exit(status: int = 0) -> None:
     """Kill the current process without fully cleaning up.